Pushed object creation and application into the engine
